mark: pallet - fix the build error (#63)
```
  --> pallets/mark/src/lib.rs:169:77
    |
169 |                 ensure!(<pallet_delegation::Module<T>>::is_delegating(&sender, &del_id, max_depth)?, Error::<T>::UnauthorizedRestore);
    |                                                                                         ^^^^^^^^^ expected `u32`, found `u64`
    |
help: you can convert a `u64` to a `u32` and panic if the converted value doesn't fit
    |
169 |                 ensure!(<pallet_delegation::Module<T>>::is_delegating(&sender, &del_id, max_depth.try_into().unwrap())?, Error::<T>::UnauthorizedRestore);
    |                                                                                         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

+++ pallets/mark/src/lib.rs
@@ -151,7 +151,7 @@ decl_module! {
 		/// where, origin is the signed sender account,
 		/// mark_hash is the revoked #MARK.
 		#[weight = 10_000_000]
-		pub fn restore(origin, mark_hash: T::Hash, max_depth: u64) -> DispatchResult {
+		pub fn restore(origin, mark_hash: T::Hash, max_depth: u32) -> DispatchResult {
 			// origin of the transaction needs to be a signed sender account
 			let sender = ensure_signed(origin)?;
